+++ app/build.gradle
@@ -21,7 +21,8 @@ android {
         }
         debug {
             buildConfigField "String", "BASE_URL", "\"http://worksample-api.herokuapp.com\""
-//            buildConfigField "String", "SAY_WEATHER_API_KEY", "\"${System.getenv(SAY_WEATHER_API_KEY)}\""
+            buildConfigField "String", "SAY_WEATHER_API_KEY", SAY_WEATHER_API_KEY
+            // in <home>/.gradle/gradle.properties
         }
     }
     sourceSets {
@@ -56,6 +57,7 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.support:appcompat-v7:27.1.0'
     implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'com.android.support:support-v4:27.1.1'
     testImplementation 'junit:junit:4.12'
     api "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     //Retrofit
