+++ app/src/main/res/values/colors.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#000000</color>
-    <color name="colorPrimaryDark">#000</color>
-    <color name="colorAccent">#FF4081</color>
+    <color name="colorPrimary">#5C5C5C</color>
+    <color name="colorPrimaryDark">#2D2D2D</color>
+    <color name="colorAccent">#FF8C18</color>
     <color name="white">#FFFFFF</color>
+    <color name="paleWhite">#F8F8F8</color>
 
     <color name="colorMorning">#eab98b</color>
     <color name="colorDay">#e59569</color>
